+++ activity_tracker/src/main/java/de/rwth/dbis/acis/activitytracker/service/dal/helpers/PaginationResult.java
@@ -4,6 +4,7 @@ package de.rwth.dbis.acis.activitytracker.service.dal.helpers;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import de.rwth.dbis.acis.activitytracker.service.dal.entities.EntityBase;
 
 import java.util.List;
@@ -40,8 +41,10 @@ public class PaginationResult<T extends EntityBase> {
     }
 
     public String toJSON() throws JsonProcessingException {
-        return new ObjectMapper().setSerializationInclusion(JsonInclude.Include.NON_NULL).writeValueAsString(this.getElements());
+        return new ObjectMapper()
+                .registerModule(new JavaTimeModule())
+                .setSerializationInclusion(JsonInclude.Include.NON_NULL)
+                .writeValueAsString(this.getElements());
     }
 
 }
-