+++ CHANGELOG.md
@@ -17,6 +17,11 @@ Releases prior to v0.6.0 are only documented on the [GitHub Release Page](https:
 - Changed buildsystem to gradle [#43](https://github.com/rwth-acis/las2peer-ActivityTracker/pull/43)
 - Automatically generate jooq code from migration files at build time [#43](https://github.com/rwth-acis/las2peer-ActivityTracker/pull/43)
 - Move to multi-staged docker build for slimmer target image [#43](https://github.com/rwth-acis/las2peer-ActivityTracker/pull/43)
+- Replaced vtor with Java Bean Validation [#44](https://github.com/rwth-acis/las2peer-ActivityTracker/pull/44)
+
+### Removed
+
+- `additionalObject` filtering capabilities in `GET /activities` endpoint [#44](https://github.com/rwth-acis/las2peer-ActivityTracker/pull/44)
 
 ## [0.6.0] - 2019-02-19